bsend: fix 2 bugs, option fail and segfault
bsend command was broken due to a confusion in the socket ids
First, set_bool_opt was passed the file descriptor instead of the
socket table index. Now the code sets the socket options to suppress
user interaction. Also, sendto_call was receiving a string instead
of the expected argv string array failing in segfault. Now we pass
the correct options to prevent prompting the user for the socket fd
since it is known.