Fix SIGINT signal being taken into account only once on Linux
On Linux, setjmp() does not save the process signal mask, which means
that the SIGINT signal will be blocked after the first run of the SIGINT
handler.

On BSD, setjmp() *does* save the process signal mask as expected.

Use sigsetjmp() and siglongjmp() instead, which are portable and have the
same semantic on both BSD and Linux.