Use a portable method to setup signal handlers
As described in signal(2), using this function to setup signal handlers is
not recommended, because it is not portable.

Switch to using sigaction() instead, with the SA_RESTART flag.  This
matches the behaviour of signal() on BSD.

This further allows to simplify signal handlers, because there is now no
need to re-setup signal handlers after each signal.  This was only needed
when signal() uses the System V semantic, which may be the case on Linux
in certain circumstances according to signal(2).